pyproject.toml
[project]
name = "llamacloud-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["llama-index-llms-openai>=0.3.28","llama-index-indices-managed-llama-cloud>=0.6.9","mcp[cli]>=1.6.0","python-dotenv>=1.1.0","llama-index-tools-mcp>=0.1.0","llama-cloud-services"]
